Añadidas las funciones de crear y deshabilitar producto, y cambiado el tipo de datos de precio

+++ src/main/java/com/example/pokeplushback/Dto/ItemCarritoDTO.java
@@ -2,10 +2,12 @@ package com.example.pokeplushback.Dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ItemCarritoDTO {
     private Integer id;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
     private Integer cantidad;
     private Integer idCarrito;
 }
